Removes trailing decimals and edit form of posts
Firstly, this removes the trailing decimals from the view of a post by parsing them before rendering. Secondly, also replaces parseInt with parseFloat.

+++ app/main/posts/modify/post-editor.directive.js
@@ -157,7 +157,8 @@ function PostEditorController(
                             $scope.post.values[attr.key] = [null];
                         }
                     }  else if (attr.input === 'number') {
-                        $scope.post.values[attr.key] = [parseInt(attr.default)];
+                        // parseFloat to handle decimals too.
+                        $scope.post.values[attr.key] = [parseFloat(attr.default)];
                     } else if (attr.input === 'date' || attr.input === 'datetime') {
                         $scope.post.values[attr.key] = attr.default ? [new Date(attr.default)] : [new Date()];
                     } else {
